Pull up AsyncPipeline to the core API
