+++ dsp/dsp-api/src/main/java/org/eclipse/dataspacetck/dsp/system/api/pipeline/ProviderNegotiationPipeline.java
@@ -16,14 +16,15 @@
 package org.eclipse.dataspacetck.dsp.system.api.pipeline;
 
 import org.eclipse.dataspacetck.dsp.system.api.statemachine.ContractNegotiation;
+import org.eclipse.dataspacetck.dsp.system.api.statemachine.ContractNegotiation.State;
 
 import java.util.Map;
 import java.util.function.Consumer;
 import java.util.function.Function;
 
 /**
- * A pipeline that drives message interactions with a provider connector under test. Uses a TCK consumer connector to interact with
- * the provider connector being verified.
+ * Constructs a contract negotiation with a provider connector under test. Uses a TCK consumer connector to interact
+ * with the provider connector being verified.
  */
 public interface ProviderNegotiationPipeline extends NegotiationPipeline<ProviderNegotiationPipeline> {
 
@@ -51,18 +52,13 @@ public interface ProviderNegotiationPipeline extends NegotiationPipeline<Provide
 
     ProviderNegotiationPipeline expectTermination();
 
-    ProviderNegotiationPipeline then(Runnable runnable);
-
-    @SuppressWarnings("unused")
-    ProviderNegotiationPipeline thenVerify(Runnable runnable);
-
     @SuppressWarnings("unused")
     ProviderNegotiationPipeline thenVerifyNegotiation(Consumer<ContractNegotiation> consumer);
 
     @SuppressWarnings("unused")
-    ProviderNegotiationPipeline thenVerifyState(ContractNegotiation.State state);
+    ProviderNegotiationPipeline thenVerifyState(State state);
 
-    ProviderNegotiationPipeline thenVerifyProviderState(ContractNegotiation.State state);
+    ProviderNegotiationPipeline thenVerifyProviderState(State state);
 
 
 }
